@@ -20,39 +20,51 @@
     "url": "git://github.com/veg/hyphy-vision.git"
   },
   "main": "dist/hyphyvision.js",
-  "version": "1.0.0",
+  "version": "1.1.0-alpha.6",
   "dependencies": {
     "bootstrap": "3.x",
+    "chi-squared": "^1.1.0",
     "crossfilter": "1.x",
+    "csvexport": "^1.0.2",
     "d3": "3.x",
+    "d3-save-svg": "0.0.2",
     "dc": "1.x",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^2.0.0",
     "font-awesome": "4.x",
     "font-awesome-webpack": "0.0.4",
     "immutable": "3.8.1",
+    "in-browser-download": "^1.0.0",
     "jquery": "1.x",
     "jquery-ui": "1.x",
     "jquery-ui-bundle": "^1.11.4",
-    "less": "^2.7.1",
+    "jquery-ui-dist": "^1.12.1",
     "lodash": "3.x",
     "lodash.clonedeep": "^4.5.0",
-    "phylotree": "0.1.4",
+    "phylotree": "0.1.6",
+    "prop-types": "^15.5.10",
     "react": "15.x.x",
-    "react-dom": "^15.4.1"
+    "react-copy-to-clipboard": "^5.0.0",
+    "react-dom": "^15.4.1",
+    "save-svg-as-png": "^1.2.0",
+    "underscore": "^1.8.3"
   },
   "devDependencies": {
     "babel-core": "^6.17.0",
+    "babel-eslint": "^7.2.3",
     "babel-jest": "^19.0.0",
-    "babel-loader": "^6.2.5",
+    "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-react-hmre": "^1.1.1",
-    "bootstrap-webpack": "0.0.5",
+    "bootstrap-webpack": "0.0.x",
     "css-loader": "^0.25.0",
-    "eslint": "0.21.x",
+    "enzyme": "^2.9.1",
+    "eslint": "^4.0.0",
+    "eslint-loader": "^1.8.0",
+    "eslint-plugin-react": "^7.1.0",
     "exports-loader": "^0.6.3",
     "expose-loader": "^0.7.1",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^2.0.0",
     "file-loader": "^0.9.0",
     "gulp": "^3.9.1",
     "gulp-concat": "^2.6.0",
@@ -63,16 +75,23 @@
     "gulp-watch": "^4.3.8",
     "imports-loader": "^0.6.5",
     "jest": "^19.0.2",
+    "less": "^2.7.2",
+    "less-loader": "^4.0.4",
+    "node-sass": "^4.5.3",
     "npm-version": "^1.1.0",
+    "react-addons-test-utils": "^15.6.0",
     "react-lazy-render": "0.3.0",
     "react-test-renderer": "^15.5.4",
+    "sass-loader": "^6.0.6",
     "style-loader": "^0.13.1",
     "uglify-js": "2.4.x",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.2"
+    "webpack": "^2.0.0",
+    "webpack-eslint-plugin": "^1.2.0"
   },
   "scripts": {
-    "lint": "eslint dist/hyphyvision.js"
+    "lint": "eslint dist/hyphyvision.js",
+    "jest": "node ./node_modules/jest-cli/bin/jest.js"
   },
   "license": "MIT",
   "files": [
