@@ -1,8 +1,10 @@
-var root = this;
-var datamonkey = function () {};
+var $ = require("jquery");
 
-if (typeof exports !== 'undefined') {
-  if (typeof module !== 'undefined' && module.exports) {
+var root = this;
+var datamonkey = function() {};
+
+if (typeof exports !== "undefined") {
+  if (typeof module !== "undefined" && module.exports) {
     exports = module.exports = datamonkey;
   }
   exports.datamonkey = datamonkey;
@@ -10,41 +12,46 @@ if (typeof exports !== 'undefined') {
   root.datamonkey = datamonkey;
 }
 
-datamonkey.errorModal = function (msg) {
-  $('#modal-error-msg').text(msg);
-  $('#errorModal').modal();
+datamonkey.errorModal = function(msg) {
+  $("#modal-error-msg").text(msg);
+  $("#errorModal").modal();
 };
 
 function b64toBlob(b64, onsuccess, onerror) {
-    var img = new Image();
 
-    img.onerror = onerror;
+  var img = new Image();
 
-    img.onload = function onload() {
-        var canvas = document.getElementById("hyphy-chart-canvas");
-        canvas.width = img.width;
-        canvas.height = img.height;
+  img.onerror = onerror;
 
-        var ctx = canvas.getContext('2d');
-      	ctx.fillStyle = "#FFFFFF";
-				ctx.fillRect(0,0,canvas.width,canvas.height);
-        ctx.drawImage(img, 0, 0, canvas.width, canvas.height);
+  img.onload = function onload() {
+    var canvas = document.getElementById("hyphy-chart-canvas");
+    canvas.width = img.width;
+    canvas.height = img.height;
 
-        canvas.toBlob(onsuccess);
-    };
+    var ctx = canvas.getContext("2d");
+    ctx.fillStyle = "#FFFFFF";
+    ctx.fillRect(0, 0, canvas.width, canvas.height);
+    ctx.drawImage(img, 0, 0, canvas.width, canvas.height);
 
-    img.src = b64;
+    canvas.toBlob(onsuccess);
+  };
+
+  img.src = b64;
+
 }
-
 
 datamonkey.export_csv_button = function(data) {
   data = d3.csv.format(data);
   if (data !== null) {
-    var pom = document.createElement('a');
-    pom.setAttribute('href', 'data:text/csv;charset=utf-8,' + encodeURIComponent(data));
-    pom.setAttribute('download', 'export.csv');
-    pom.className = 'btn btn-default btn-sm';
-    pom.innerHTML = '<span class="glyphicon glyphicon-floppy-save"></span> Download CSV';
+    var pom = document.createElement("a");
+    pom.setAttribute(
+      "href",
+      "data:text/csv;charset=utf-8," + encodeURIComponent(data)
+    );
+    pom.setAttribute("download", "export.csv");
+    pom.className = "btn btn-default btn-sm";
+    pom.innerHTML =
+      '<span class="glyphicon glyphicon-floppy-save"></span> Download CSV';
     $("body").append(pom);
     pom.click();
     pom.remove();
@@ -52,7 +59,6 @@ datamonkey.export_csv_button = function(data) {
 };
 
 datamonkey.save_image = function(type, container) {
-
   var prefix = {
     xmlns: "http://www.w3.org/2000/xmlns/",
     xlink: "http://www.w3.org/1999/xlink",
@@ -60,7 +66,6 @@ datamonkey.save_image = function(type, container) {
   };
 
   function get_styles(doc) {
-
     function process_stylesheet(ss) {
       try {
         if (ss.cssRules) {
@@ -80,12 +85,12 @@ datamonkey.save_image = function(type, container) {
           }
         }
       } catch (e) {
-        console.log('Could not process stylesheet : ' + ss);
+        //console.log("Could not process stylesheet : " + ss);
       }
     }
 
     var styles = "",
-        styleSheets = doc.styleSheets;
+      styleSheets = doc.styleSheets;
 
     if (styleSheets) {
       for (var i = 0; i < styleSheets.length; i++) {
@@ -94,35 +99,7 @@ datamonkey.save_image = function(type, container) {
     }
 
     return styles;
-
   }
-
-  var convert_svg_to_png = function(image_string) {
-
-    var image = document.getElementById("hyphy-chart-image");
-
-    image.onload = function() {
-
-      var canvas = document.getElementById("hyphy-chart-canvas");
-      canvas.width = image.width;
-      canvas.height = image.height;
-      var context = canvas.getContext("2d");
-      context.fillStyle = "#FFFFFF";
-      context.fillRect(0,0,image.width,image.height);
-      context.drawImage(image, 0, 0);
-      var img = canvas.toDataURL("image/png");
-      var pom = document.createElement('a');
-      pom.setAttribute('download', 'image.png');
-      pom.href = canvas.toDataURL("image/png");     
-      $("body").append(pom);
-      pom.click();
-      pom.remove();
-
-    };
-
-    image.src = image_string;
-
-  };
 
   var svg = $(container).find("svg")[0];
   if (!svg) {
@@ -134,12 +111,11 @@ datamonkey.save_image = function(type, container) {
   svg.setAttribute("version", "1.1");
 
   var defsEl = document.createElement("defs");
-  svg.insertBefore(defsEl, svg.firstChild); 
+  svg.insertBefore(defsEl, svg.firstChild);
 
   var styleEl = document.createElement("style");
   defsEl.appendChild(styleEl);
   styleEl.setAttribute("type", "text/css");
-
 
   // removing attributes so they aren't doubled up
   svg.removeAttribute("xmlns");
@@ -154,80 +130,82 @@ datamonkey.save_image = function(type, container) {
     svg.setAttributeNS(prefix.xmlns, "xmlns:xlink", prefix.xlink);
   }
 
-  var source = (new XMLSerializer()).serializeToString(svg).replace('</style>', '<![CDATA[' + styles + ']]></style>');
-  var rect = svg.getBoundingClientRect();
-  var doctype = '<?xml version="1.0" standalone="no"?><!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">';
+  var source = new XMLSerializer()
+    .serializeToString(svg)
+    .replace("</style>", "<![CDATA[" + styles + "]]></style>");
+  var doctype =
+    '<?xml version="1.0" standalone="no"?><!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">';
   var to_download = [doctype + source];
-  var image_string = 'data:image/svg+xml;base66,' + encodeURIComponent(to_download);
+  var image_string =
+    "data:image/svg+xml;base66," + encodeURIComponent(to_download);
 
-  if(type == "png") {
-		b64toBlob(image_string,
-				function(blob) {
-
-						var url = window.URL.createObjectURL(blob);
-						var pom = document.createElement('a');
-						pom.setAttribute('download', 'image.png');
-						pom.setAttribute('href', url);
-						$("body").append(pom);
-						pom.click();
-						pom.remove();
-
-				}, function(error) {
-						// handle error
-				});
+  if (type == "png") {
+    b64toBlob(
+      image_string,
+      function(blob) {
+        var url = window.URL.createObjectURL(blob);
+        var pom = document.createElement("a");
+        pom.setAttribute("download", "image.png");
+        pom.setAttribute("href", url);
+        $("body").append(pom);
+        pom.click();
+        pom.remove();
+      },
+      function(error) {
+        // handle error
+      }
+    );
   } else {
-    var pom = document.createElement('a');
-    pom.setAttribute('download', 'image.svg');
-    pom.setAttribute('href', image_string);
+    var pom = document.createElement("a");
+    pom.setAttribute("download", "image.svg");
+    pom.setAttribute("href", image_string);
     $("body").append(pom);
     pom.click();
     pom.remove();
   }
-
 };
 
-datamonkey.validate_date = function () {
-
+datamonkey.validate_date = function() {
   // Check that it is not empty
-  if($(this).val().length === 0) {
-    $(this).next('.help-block').remove();
-    $(this).parent().removeClass('has-success');
-    $(this).parent().addClass('has-error');
+  if ($(this).val().length === 0) {
+    $(this).next(".help-block").remove();
+    $(this).parent().removeClass("has-success");
+    $(this).parent().addClass("has-error");
 
-    jQuery('<span/>', {
-          class: 'help-block',
-          text : 'Field is empty'
-      }).insertAfter($(this));
+    jQuery("<span/>", {
+      class: "help-block",
+      text: "Field is empty"
+    }).insertAfter($(this));
+  } else if (isNaN(Date.parse($(this).val()))) {
+    $(this).next(".help-block").remove();
+    $(this).parent().removeClass("has-success");
+    $(this).parent().addClass("has-error");
 
-  } else if(isNaN(Date.parse($(this).val()))) {
-    $(this).next('.help-block').remove();
-    $(this).parent().removeClass('has-success');
-    $(this).parent().addClass('has-error');
-
-    jQuery('<span/>', {
-          class: 'help-block',
-          text : 'Date format should be in the format YYYY-mm-dd'
-      }).insertAfter($(this));
-
+    jQuery("<span/>", {
+      class: "help-block",
+      text: "Date format should be in the format YYYY-mm-dd"
+    }).insertAfter($(this));
   } else {
-    $(this).parent().removeClass('has-error');
-    $(this).parent().addClass('has-success');
-    $(this).next('.help-block').remove();
+    $(this).parent().removeClass("has-error");
+    $(this).parent().addClass("has-success");
+    $(this).next(".help-block").remove();
   }
-
 };
 
-$( document ).ready( function () {
-  $(function () {$('[data-toggle="tooltip"]').tooltip();});
-  $('#datamonkey-header').collapse ();
-  
+$(document).ready(function() {
+  $(function() {
+    $('[data-toggle="tooltip"]').tooltip();
+  });
+  $("#datamonkey-header").collapse();
+
   var initial_padding = $("body").css("padding-top");
-  
-  $("#collapse_nav_bar").on ("click", function (e) {
-    $('#datamonkey-header').collapse ('toggle');
-    $(this).find ("i").toggleClass ("fa-times-circle fa-eye");
-    var new_padding =  $("body").css("padding-top") == initial_padding ? "5px" : initial_padding;
-    d3.select ("body").transition ().style ("padding-top", new_padding);
+
+  $("#collapse_nav_bar").on("click", function(e) {
+    $("#datamonkey-header").collapse("toggle");
+    $(this).find("i").toggleClass("fa-times-circle fa-eye");
+    var new_padding = $("body").css("padding-top") == initial_padding
+      ? "5px"
+      : initial_padding;
+    d3.select("body").transition().style("padding-top", new_padding);
   });
 });
-
