@@ -1,31 +1,17 @@
-var React = require('react'),
-    ReactDOM = require('react-dom');
+var React = require("react"),
+  _ = require('underscore');
 
-var InputInfo = React.createClass({
-
-  getInitialState(){
-    return {
-      input_data: {}
-    };
-  },
-
-  componentWillReceiveProps(nextProps){
-    this.setState({
-      input_data: nextProps.input_data
-    });
-  },
-
-  render(){
-    return(
-      <div id="input-info">
-        <h5 className="hyphy-highlight">INPUT DATA</h5><br/>
-        <p>{this.state.input_data.filename}</p><br/>
-        <p><span className="hyphy-highlight">{this.state.input_data.sequences}</span> sequences</p><br/>
-        <p><span className="hyphy-highlight">{this.state.input_data.sites}</span> sites</p><br/>
-      </div>
-    )
-  }
-
-});
+function InputInfo(props){
+  if(!props.input_data) return <div></div>;
+  var is_full_path = props.input_data['file name'].indexOf('/') != -1,
+    filename = is_full_path ? _.last(props.input_data['file name'].split('/')) : props.input_data['file name'];
+  return (<div id="input-info">
+    <span className="hyphy-highlight">INPUT DATA</span> <span className="divider">|</span>
+    <a href="#">{filename}</a> <span className="divider">|</span>
+    <span className="hyphy-highlight">{props.input_data['number of sequences']}</span> sequences <span className="divider">|</span>
+    <span className="hyphy-highlight">{props.input_data['number of sites']}</span> sites
+  </div>);
+}
 
 module.exports.InputInfo = InputInfo;
+
