@@ -1,91 +1,129 @@
-var React = require('react');
+var React = require("react");
+
 
 var Hamburger = React.createClass({
-  render: function(){
+  render: function() {
     return (
-      <button type="button" className="navbar-toggle" data-toggle="collapse" data-target="#navbar-collapse-1">
+      <button
+        type="button"
+        className="navbar-toggle"
+        data-toggle="collapse"
+        data-target="#navbar-collapse-1"
+      >
         <span className="sr-only">Toggle navigation</span>
-        <span className="icon-bar"></span>
-        <span className="icon-bar"></span>
-        <span className="icon-bar"></span>
+        <span className="icon-bar" />
+        <span className="icon-bar" />
+        <span className="icon-bar" />
       </button>
-    )
+    );
   }
 });
 
 var Methods = React.createClass({
-  render: function(){
+  render: function() {
     return (
-      <ul className="nav nav-pills navbar-right" id='navigation_buttons'>
-        <li className="active">
-          <a href="#" role="tab">absrel</a>
-        </li>
-        <li>
-          <a href="../relax" role="tab">relax</a>
-        </li>
-        <li>
-          <a href="../busted" role="tab">busted</a>
-        </li>
-        <li>
-          <a href="../fade" role="tab">fade</a>
-        </li>
-        <li>
-          <a href="../slac" role="tab">slac</a>
-        </li>
-      </ul>
-    )
+      <div className="dropdown">
+        <button
+          className="btn btn-primary dropdown-toggle"
+          type="button"
+          data-toggle="dropdown"
+        >
+          Tools
+          <span className="caret" />
+        </button>
+        <ul className="dropdown-menu">
+          <li>
+            <a href="../absrel">aBSREL</a>
+          </li>
+          <li>
+            <a href="../relax">RELAX</a>
+          </li>
+          <li>
+            <a href="../busted">BUSTED</a>
+          </li>
+          <li>
+            <a href="../fade">FADE</a>
+          </li>
+          <li>
+            <a href="../slac">SLAC</a>
+          </li>
+        </ul>
+      </div>
+    );
   }
 });
 
 var NavBar = React.createClass({
   componentDidMount: function() {
     // Corrects navbar offset when clicking anchor hash
-    var shiftWindow = function() { scrollBy(0, -50) };
+    var shiftWindow = function() {
+      scrollBy(0, -50);
+    };
     if (location.hash) shiftWindow();
     window.addEventListener("hashchange", shiftWindow);
-
   },
-  render: function(){
+  render: function() {
+    var self = this,
+      input_style = {
+      position: "absolute",
+      top: 0,
+      right: 0,
+      minWidth: "100%",
+      minHeight: "100%",
+      fontSize: "100px",
+      textAlign: "right",
+      filter: "alpha(opacity=0)",
+      opacity: 0,
+      outline: "none",
+      background: "white",
+      cursor: "inherit",
+      display: "block"
+    };
     return (
-      <nav className="navbar navbar-default navbar-fixed-top" role="navigation">
-        <div className="container-fluid">
+      <nav className="navbar navbar-default navbar-fixed-top main-nav" role="navigation">
+        <div className="container">
           <div className="row">
-
-            <div className="col-sm-1"></div>
-
-            <div className="col-sm-10">
-              <a href='#'><img id="hyphy-logo" src="../../images/hyphy-logo.svg" /></a>
+            <div className="col-sm-12">
+              <a href="#">
+                <img id="hyphy-logo" src="../../images/hyphy-logo.svg" />
+              </a>
               <div className="navbar-header">
                 <Hamburger />
-                <a className="navbar-brand logo" href="#">HYPHY VISION</a>
               </div>
 
               <div className="collapse navbar-collapse" id="navbar-collapse-1">
-
                 <ul className="nav navbar-nav">
-                  <li className="dropdown">
-                    <a href="#" className="dropdown-toggle" id="datamonkey-absrel-toggle-here" data-toggle="dropdown">Load file<b className="caret"></b></a>
-                    <ul className="dropdown-menu">
-                      <li>
-                        <input type="file" id="datamonkey-absrel-json-file" />
-                      </li>
-                    </ul>
-                  </li>
+                  <a
+                    href="#"
+                    className="nav-button"
+                    role="button"
+                    style={{ position: "relative", overflow: "hidden" }}
+                  >
+                    <input 
+                      type="file"
+                      style={input_style}
+                      id="dm-file"
+                      onChange={self.props.onFileChange}
+                    />
+                    Load
+                  </a>
+                  <a
+                    href="#"
+                    className="nav-button"
+                    role="button"
+                    style={{ display: "none" }}
+                  >
+                    Export
+                  </a>
                 </ul>
-
-              <Methods />
-                
+                <Methods />
               </div>
             </div>
-
-            <div className="col-sm-1"></div>
-
           </div>
         </div>
       </nav>
-    )
+    );
   }
 });
 
 module.exports.NavBar = NavBar;
-
